tickets/serializers.py
from rest_framework import serializers
from .models import Reserva

class ReservaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Reserva
        fields = ['id', 'nombre_cliente', 'cantidad_tickets', 'estado', 'evento']
