pyproject.toml
[tool.poetry]
name = "reservas-ticket"
version = "0.1.0"
description = ""
authors = ["Freddy Morales <freddy_alexander@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
mysqlclient = "^2.2.7"
celery = "^5.4.0"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
